+++ contract/src/test/java/org/obolibrary/obo2owl/RoundTripCAROTestCase.java
@@ -3,7 +3,7 @@ package org.obolibrary.obo2owl;
 import org.junit.Test;
 
 @SuppressWarnings("javadoc")
-public class RoundTripCAROTest extends RoundTripTest {
+public class RoundTripCAROTestCase extends RoundTripTest {
 
     @Test
     public void testRoundTrip() throws Exception {
