+++ contract/src/test/java/org/obolibrary/obo2owl/RoundTripCardinality2TestCase.java
@@ -3,7 +3,7 @@ package org.obolibrary.obo2owl;
 import org.junit.Test;
 
 @SuppressWarnings("javadoc")
-public class RoundTripCardinality extends RoundTripTest {
+public class RoundTripCardinality2TestCase extends RoundTripTest {
 
     @Test
     public void roundTripCardinality() throws Exception {
