+++ contract/src/test/java/org/obolibrary/obo2owl/RoundTripEquivalentToChainTestCase.java
@@ -3,7 +3,7 @@ package org.obolibrary.obo2owl;
 import org.junit.Test;
 
 @SuppressWarnings("javadoc")
-public class RoundTripEquivalentToChain extends RoundTripTest {
+public class RoundTripEquivalentToChainTestCase extends RoundTripTest {
 
     @Test
     public void testTrailingQualifiers() throws Exception {
