+++ contract/src/test/java/org/obolibrary/obo2owl/RoundTripOboAnnotationsTestCase.java
@@ -18,7 +18,7 @@ import org.semanticweb.owlapi.model.OWLOntology;
 import org.semanticweb.owlapi.model.OWLSubClassOfAxiom;
 
 @SuppressWarnings("javadoc")
-public class RoundTripOboAnnotations extends RoundTripTest {
+public class RoundTripOboAnnotationsTestCase extends RoundTripTest {
 
     @Test
     public void testIsInferredAnnotation() throws Exception {
