+++ contract/src/test/java/org/semanticweb/owlapi/util/FormatsMimeTestCase.java
@@ -32,7 +32,7 @@ import org.semanticweb.owlapi.io.OWLParserFactory;
  */
 @SuppressWarnings("javadoc")
 @RunWith(Parameterized.class)
-public class FormatsMimeTestcase {
+public class FormatsMimeTestCase {
     @Parameters(name = "{1}")
     public static List<Object[]> params() {
         return Arrays.asList(
@@ -41,15 +41,15 @@ public class FormatsMimeTestcase {
             new Object[] { new org.semanticweb.owlapi.krss2.parser.KRSS2OWLParserFactory(),                               "KRSS2 Syntax",           null,                       Collections.emptyList()},
             new Object[] { new org.semanticweb.owlapi.dlsyntax.parser.DLSyntaxOWLParserFactory(),                         "DL Syntax Format",       null,                       Collections.emptyList()},
             new Object[] { new org.semanticweb.owlapi.rdf.rdfxml.parser.RDFXMLParserFactory(),                            "RDF/XML Syntax",         "application/rdf+xml",      Arrays.asList("application/rdf+xml",    "application/xml",  "text/xml")},
-            new Object[] { new org.semanticweb.owlapi.rio.RioRDFXMLParserFactory(),                                       "RDF/XML",                "application/rdf+xml",      Arrays.asList("application/rdf+xml",    "application/xml",  "text/xml")},
-            new Object[] { new org.semanticweb.owlapi.rio.RioNQuadsParserFactory(),                                       "N-Quads",                "application/n-quads",      Arrays.asList("application/n-quads",    "text/x-nquads",    "text/nquads")},
+            new Object[] { new org.semanticweb.owlapi.rio.RioRDFXMLParserFactory(),                                       "RDF/XML",                "application/rdf+xml",      Arrays.asList("application/rdf+xml",    "application/xml")},
+            new Object[] { new org.semanticweb.owlapi.rio.RioNQuadsParserFactory(),                                       "N-Quads",                "text/x-nquads",            Arrays.asList("text/x-nquads")},
             new Object[] { new org.semanticweb.owlapi.rio.RioRDFaParserFactory(),                                         "RDFa",                   "application/xhtml+xml",    Arrays.asList("application/xhtml+xml",  "application/html", "text/html")},
             new Object[] { new org.semanticweb.owlapi.rdf.turtle.parser.TurtleOntologyParserFactory(),                    "Turtle Syntax",          "text/turtle",              Arrays.asList("text/turtle",            "application/x-turtle")},
             new Object[] { new org.semanticweb.owlapi.rio.RioTurtleParserFactory(),                                       "Turtle",                 "text/turtle",              Arrays.asList("text/turtle",            "application/x-turtle")},
             new Object[] { new org.semanticweb.owlapi.rio.RioN3ParserFactory(),                                           "N3",                     "text/n3",                  Arrays.asList("text/n3",                "text/rdf+n3")},
             new Object[] { new org.semanticweb.owlapi.owlxml.parser.OWLXMLParserFactory(),                                "OWL/XML Syntax",         "application/owl+xml",      Arrays.asList("application/owl+xml",    "text/xml")},
-            new Object[] { new org.semanticweb.owlapi.rio.RioNTriplesParserFactory(),                                     "N-Triples",              "application/n-triples",    Arrays.asList("application/n-triples",  "text/plain")},
-            new Object[] { new org.semanticweb.owlapi.rio.RioTrigParserFactory(),                                         "TriG",                   "application/trig",         Arrays.asList("application/trig",       "application/x-trig")},
+            new Object[] { new org.semanticweb.owlapi.rio.RioNTriplesParserFactory(),                                     "N-Triples",              "text/plain",               Arrays.asList("text/plain")},
+            new Object[] { new org.semanticweb.owlapi.rio.RioTrigParserFactory(),                                         "TriG",                   "application/x-trig",       Arrays.asList("application/x-trig")},
             new Object[] { new org.semanticweb.owlapi.manchestersyntax.parser.ManchesterOWLSyntaxOntologyParserFactory(), "Manchester OWL Syntax",  "text/owl-manchester",      Arrays.asList("text/owl-manchester")},
             new Object[] { new org.semanticweb.owlapi.functional.parser.OWLFunctionalSyntaxOWLParserFactory(),            "OWL Functional Syntax",  "text/owl-functional",      Arrays.asList("text/owl-functional")},
             new Object[] { new org.semanticweb.owlapi.rio.RioBinaryRdfParserFactory(),                                    "BinaryRDF",              "application/x-binary-rdf", Arrays.asList("application/x-binary-rdf")},
